Move `GridWrapper` to `components` folder
