+++ src/components/domain/post/blog/BlogPosts.tsx
@@ -5,7 +5,7 @@ import NextLink from '../../../core/NextLink'
 import Flex from '../../../core/Flex'
 import Heading from '../../../core/Heading'
 import Text from '../../../core/Text'
-import GridWrapper from '../GridWrapper'
+import GridWrapper from '../../../GridWrapper'
 import PostPreviewCard from '../PostPreviewCard'
 import {routes} from '../../../../routes'
 import PostTag from '../../buttons/PostTag'
