+++ src/components/domain/post/podcast/Podcasts.tsx
@@ -5,7 +5,7 @@ import NextLink from '../../../core/NextLink'
 import Flex from '../../../core/Flex'
 import Heading from '../../../core/Heading'
 import Text from '../../../core/Text'
-import GridWrapper from '../GridWrapper'
+import GridWrapper from '../../../GridWrapper'
 import PostPreviewCard from '../PostPreviewCard'
 import {routes} from '../../../../routes'
 import {createMapOfPostsByTag, getTopNTags} from '../postUtils'
